requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
openpyxl>=3.1.0
rapidfuzz>=3.10.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pytest>=8.0.0
ruff>=0.9.0
mcp>=1.0.0
